Update story-viz to save selection of images
Use mongodb to store the latest selection of images. This will allow
us to publish the story by simply showing the selections for now.  We
may eventually need to actually publish the story with only the selected
data if we find that there is some data that we don't want the public
to be able to see.