Add StoryList as front page and local FastAPI
Create a landing page for story-viz so that we can select a story and edit it.
This landing page now has the publish button on the landing page instead of
on the story page. This needs review as we develop the app.

Also add a Login page that fetches a JWT token from the backend so that we can
safely expose story-viz to the web.  This will eventually enable using story-viz
as the basis of a public interface for creating and sharing stories.